requirements.txt
pandas
sqlalchemy
statsmodels==0.14.0
matplotlib==3.7.0
seaborn==0.12.0
mlflow>=2.8.0
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic>=2.4.0
requests>=2.31.0
python-multipart
huggingface-hub
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
scikit-learn
joblib